// src/types/enum.ts
export enum AppEvents {
    ItemsChanged = 'items:changed',
    CardSelect = 'card:select',
    CardAddToBasket = 'card:addToBasket',
    BasketOpen = 'basket:open',
    ModalClose = 'modal:close',
    BasketDelete = 'basket:delete',
    BasketOrder = 'basket:order',
    OrderFormErrorsChange = 'orderFormErrors:change',
    ContactsFormErrorsChange = 'contactsFormErrors:change',
    OrderInputChange = 'orderInput:change',
    OrderSubmit = 'order:submit',
    OrderSuccess = 'order:success',
    ContactsSubmit = 'contacts:submit'
}